+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- [added] `messaging.Aps` type now supports critical sound in its payload.
+
 # v3.5.0
 
 - [added] `messaging.AndroidNotification` type now supports `channel_id`.
@@ -7,6 +9,7 @@
 - [fixed] Fixing error handling in FCM. The SDK now checks the key 
   `type.googleapis.com/google.firebase.fcm.v1.FcmError` to set error code.
 - [added] `messaging.ApsAlert` type now supports subtitle in its payload.
+- [added] `messaging.WebpushConfig` type now supports fcmOptions in its payload.
 
 # v3.4.0